+++ 0.2/Control/Applicative/Backwards.hs
@@ -17,11 +17,16 @@ module Control.Applicative.Backwards where
 import Prelude hiding (foldr, foldr1, foldl, foldl1)
 import Control.Applicative
 import Data.Foldable
+import Data.Functor.Classes
 import Data.Traversable
 
 -- | The same functor, but with an 'Applicative' instance that performs
 -- actions in the reverse order.
-newtype Backwards f a = Backwards { forwards :: f a }
+newtype Backwards f a = Backwards (f a)
+
+-- | Inverse of 'Backwards'.
+forwards :: Backwards f a -> f a
+forwards (Backwards x) = x
 
 -- | Derived instance.
 instance (Functor f) => Functor (Backwards f) where
@@ -49,3 +54,22 @@ instance (Foldable f) => Foldable (Backwards f) where
 instance (Traversable f) => Traversable (Backwards f) where
     traverse f (Backwards t) = fmap Backwards (traverse f t)
     sequenceA (Backwards t) = fmap Backwards (sequenceA t)
+
+
+instance (Eq1 f, Eq a) => Eq (Backwards f a) where
+    Backwards x == Backwards y = eq1 x y
+
+instance (Ord1 f, Ord a) => Ord (Backwards f a) where
+    compare (Backwards x) (Backwards y) = compare1 x y
+
+instance (Read1 f, Read a) => Read (Backwards f a) where
+    readsPrec = readsData $ readsUnary1 "Backwards" Backwards
+
+instance (Show1 f, Show a) => Show (Backwards f a) where
+    showsPrec d (Backwards x) = showsUnary1 "Backwards" d x
+
+instance Eq1 f => Eq1 (Backwards f) where eq1 = (==)
+instance Ord1 f => Ord1 (Backwards f) where compare1 = compare
+instance Read1 f => Read1 (Backwards f) where readsPrec1 = readsPrec
+instance Show1 f => Show1 (Backwards f) where showsPrec1 = showsPrec
+