+++ 0.2/Data/Functor/Reverse.hs
@@ -20,6 +20,7 @@ import Control.Applicative.Backwards
 import Prelude hiding (foldr, foldr1, foldl, foldl1)
 import Control.Applicative
 import Data.Foldable
+import Data.Functor.Classes
 import Data.Traversable
 import Data.Monoid
 
@@ -55,3 +56,21 @@ instance (Traversable f) => Traversable (Reverse f) where
         fmap Reverse . forwards $ traverse (Backwards . f) t
     sequenceA (Reverse t) =
         fmap Reverse . forwards $ sequenceA (fmap Backwards t)
+
+instance (Eq1 f, Eq a) => Eq (Reverse f a) where
+    Reverse x == Reverse y = eq1 x y
+
+instance (Ord1 f, Ord a) => Ord (Reverse f a) where
+    compare (Reverse x) (Reverse y) = compare1 x y
+
+instance (Read1 f, Read a) => Read (Reverse f a) where
+    readsPrec = readsData $ readsUnary1 "Reverse" Reverse
+
+instance (Show1 f, Show a) => Show (Reverse f a) where
+    showsPrec d (Reverse x) = showsUnary1 "Reverse" d x
+
+instance Eq1 f => Eq1 (Reverse f) where eq1 = (==)
+instance Ord1 f => Ord1 (Reverse f) where compare1 = compare
+instance Read1 f => Read1 (Reverse f) where readsPrec1 = readsPrec
+instance Show1 f => Show1 (Reverse f) where showsPrec1 = showsPrec
+